requirements.txt
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
mask-rcnn==2.1
matplotlib==3.1.3
numpy==1.18.1
pandas==0.24.1
Pillow==7.0.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
sklearn==0.0
streamlit==0.57.1
tensorboard==1.15.0
tensorboardX==2.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1


